refactor: add JS modules and SCSS
