+++ source/index.html
@@ -5,12 +5,12 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="css/style.min.css">
   <title>Virtual Keyboard</title>
 </head>
 
 <body>
-  <script src="js/keys.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
