+++ source/js/main.js
@@ -1,0 +1,1 @@
+import './data.js';
